+++ backend/src/main/java/com/bubble/backend/service/UserService.java
@@ -8,7 +8,7 @@ public interface UserService {
 
     void addUser(User user);
 
-    User getUserById(long id);
+    User getUserById(int id);
 
     List<User> getAllUsers();
