+++ backend/src/main/java/com/bubble/backend/serviceImpl/UserServiceImpl.java
@@ -20,7 +20,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserById(long id) {
+    public User getUserById(int id) {
         return userRepository.findById(id).get(0);
     }
